requirements.txt
torch>=1.13.1
torchvision>=0.14.0
torchaudio>=0.13.0
transformers>=4.30.0
datasets>=2.12.0
peft>=0.4.0
bitsandbytes>=0.39.0
accelerate>=0.20.0
evaluate>=0.4.0
sentencepiece>=0.1.99
protobuf>=3.20.0,<4.0.0
scipy>=1.10.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
pandas>=2.0.0
tqdm>=4.65.0
omegaconf>=2.3.0
safetensors>=0.3.1
opencv-python>=4.7.0
ffmpeg-python>=0.2.0
librosa>=0.10.0
soundfile>=0.12.0
pytorch-lightning>=2.0.0
wandb>=0.15.0
tensorboard>=2.12.0
hydra-core>=1.3.0
timm>=0.9.0
pyannote.audio>=2.1.0
editdistance>=0.6.0
av>=10.0.0
onnx>=1.14.0
onnxruntime>=1.14.0
pytest>=7.3.0
mypy>=1.3.0
pylint>=2.17.0
black>=23.3.0
flake8>=6.0.0
numpy>=1.21.0
pillow>=9.0.0
PyYAML>=6.0
dataclasses-json>=0.5.7

# Optional dependencies
# fairseq is optional and may cause dependency conflicts
# If you need the original AV-HuBERT implementation, try:
# fairseq>=0.12.0
